Extract TopBar into sub component
